Fixed IntelliSence putting red lines on all code
Including the boost header files makes IntelliSence assume all the DLL code is in namespace boost::python::detail and alerts that functions and enums can't be found.
Not setting the boost include path for VC will make it react to python code not being defined, but everything else will not have false errors.

The actual compilation is done by the Makefile, which is unaffected by this issue. It's purely an IDE issue.