Added CvXMLLoadUtility::GetEnum
It works like vanilla Get functions, except it uses an enum as variable argument.

It allows reading a string and converting it to an enum value in a single line where vanilla uses two lines for this.
Unlike vanilla, this one locks to look up the string in just one file and picks the file automatically based on the enum type.
This prevents the risk of reading a string from the wrong file, usually mixing up Unit and UnitClass.
It also has two assert checks to tell not just if, but also what went wrong. Vanilla has no useful asserts for this.