Greatly reduced complexity in calling CvPlayer::applyCivEffect()
Removed 3 arguments and added one. Now the arguments are:
- CivEffect as enum or const pointer to the info class (overloaded)
- bool bAdd. Default true. Set to false to remove a CivEffect from a player

The lost bool arguments are replaced with gDLL->isGameInitializing() and a private function with an extra argument.

On top of being easier to use, it reduces the risk of bugs because it's no longer possible to call and skip updating secondary cache (or screen in the future)