Fix #54: forgetting to add & when storing a get*Info() return value will now cause a compilation error
This is done by making CvInfoBase boost::noncopyable.

The problem in omitting the & is that it will create a copy of the info class instead of a reference to the same memory.
Since there is no copy operator, this will cause pointers to be copied, meaning both original and copy will point to the same memory.
When the copy goes out of scope, it will free the memory still used by the original and pointers to freed memory renders the game unstable.
This cause a lot of weird bugs, none of them pointing towards where the actual problem is because everything works until the copy goes out of scope.