#174 Show a random image each time the loading screen is on
Changed Loading into a list of Loading strings in the MainMenu xml.
Each time CvMainMenuInfo::getLoading() is called, pick one from the list to show.
The selector is based on milliseconds since system start, meaning rather than random, it loops the list a speed of 1000 images each second.
While it's technically not random at all, it certainly feels random when you click relative to that counter.