Fix: combat (game) freezes when defeating the last unit of a (European only?) player
When the last unit is defeated, the player is set to die. Game update will then kill all units and cities.
If a unit is killed (memory released) before the combat animation is done, the combat freezes.
This fix will postpone killing a player until the first game update where no units are in combat.