Fix: launchChooseProfessionPopup() can now handle when eWorkingBuilding is NO_BUILDING
This is related to multiple professions in the same building where the code assumes all professions to have a building.
The test savegame encountered this with farmers and hunters.

This resulted in choose profession showing up for outdoor professions once each turn, but for unknown reasons not for player 0.
This means it's an issue for multiplayer, but not the host.