Refactoring: Moved code for awarding father points for production of tangible yields.
From after selling to domestic market to right after calculation of net production ammounts.

This awards father points for all the produced amounts, not just for those left, after selling to the domestic market. This does not change the amount of father points awarded, as - according to xml definitions - none are awarded for the production of tangible yields at the moment (see <FatherPointInfo><YieldPoints> in CIV4FatherPointInfos.xml).