Added unicode support in text xml files
Text xml files will now be assumed to be encoded in UTF-8 if UTF8 is incided in the name.
This removes the need for html escapes and allows translations for different codepages in the same file.

No change ingame. The text is converted with iconv on load, though we use a very reduced set of iconv (only the codepages we use)