Fix: Languages not using codepage 1252 now have a better chance of not causing error messages when displaying English
Reading priority is not language itself, Default, English, Tag.
When not reading the language itself and not using codepage 1252. most characters are converted to ASCII, like á -> a.
Sure there is a loss in those conversions, but conversions only triggers when the codepage lacks the character in question,
meaning the alternative is unprintable characters. This avoids a bunch of error messages at startup in partially translated non-1252 languages.