Added ability to switch GameFont.tga according to CodePage of the selected language
One of the problems with translation support is that some characters are in GameFont.tga. Displaying those for non-western European will require another GameFont.tga
This change will allow the game to pick a GameFont.tga matching the codepage for the chosen language.
The problem is that the GameFont is loaded at startup, meaning switching the language could require a game restart (popup notification added for this)

Another issue is the need to maintain multiple GameFont files. Not ideal, but doable. It's not like we update GameFont that often anyway.

No extra GameFont added with this revision. It's just the engine to support more GameFont files. Adding files is a trivial xml edit.