Fix #258: Abandoning a colony will no longer allow setting professions without paying the yield cost
The vanilla setting CONSUME_EQUIPMENT_ON_FOUND is intended to allow pioneers to consume tools on founding.
Since we don't use this and it was the cause of this bug, I decided to just remove all references to CONSUME_EQUIPMENT_ON_FOUND.

If we ever want to consume yields on founding, we will need to make a new implementation anyway, one without such a gigantic exploit.