BUGFIX : Defensive units can now get withdraw promotion
The original code, inherited from Civ IV, prevented defensive units to get withdrawal promotions .
But we need those for trade naval units at least, for the "Merchant Navy" promotion series.

In RaR or WTP, withdrawal chances represent the ability to flee combat, regardless if the unit is the attacker or the defender. 

I don't think this changes will hurt.