CivEffects can now grant free promotions + changes to how promotions are handled in CvUnit
CvUnit now caches if a promotion is applied or not in a BoolArray.
This allows CvPlayer to grant free promotions and each unit can figure out if it should be applied or not.
Promotion data is reset on load and all promotions applied again. This makes xml changes take effect in existing games.