Moved CivEffects code from CvPlayer.h to CvPlayerCivEffect.h
CvPlayerAI now inherits CvPlayerCivEffect, which inherits CvPlayer.
Calling CivEffect is now done though CvPlayer::CivEffect(). AI() is also added for convenience.
This allows the CivEffect code to be in a file with nothing else, hence more organized.
It also allows the CivEffect cache to be private and not shared with CvPlayer.