XML assumption failures will now trigger errors instead of assert failures
This means release DLL files will now tell if the DLL has hardcoded xml data, which has later been changed in xml.
Even non-hardcoded DLLs have assumptions (like UnitAI) and will trigger this. The error message tells if a non-hardcoded DLL files the issue or if a recompile is needed.