SavegameVariableTypes is now saved using 1 or 2 bytes depending on the enum size for each class
Whenever a class saves a variable, there is a SavegameVariableTypes prefix to tell which variable is saved.
If the enum has max 255 variables (+end of class marker), this prefix is one byte, but it will automatically become 2 if the enum grows "too much".
The size for each class is saved (char array at start of savegame), allowing loading even if the enum has grown past the 255 threshold in the meantime.

While it only saves one byte each time SavegameVariableTypes is saved, it has to be multiplied by the number of variables saved and it quickly adds up.
There is no extra programing work in loading/saving data. It's all activated by CvSavegame through function overloading and as such other files can't tell the difference.
This will hopefully make it very resistant to causing bugs when adding more data to savegames in the future.