Savegames now autogenerates xml conversion tables as needed
When a file is saved, whenever a certain variable is saved, the conversion table for that type is added if it's not already added.
This eliminates the risk of bugs due to programmers not setting the list of included tables correctly.
It also makes savegames smaller in some cases because it can skip the tables for unsaved xml files even if some other savegame has to include that file.