Rewrote how team revealed route and improvement is stored in CvPlot and in savegames
What the player can see is now saved as an array of improvement,route pairs. This improves memory I/O since they are usually used together.
Store using enum types rather than shorts despite not increasing memory usage. This avoids type casting and is debugger friendly.

Saves using CvSavegame and stores teams with up to date info in a PlayerBoolArray (9 bytes in total)
Teams with some info, which is not NO_ and is out of date is stored using 4 bytes for each such team.
Previous usage was 198 bytes.
All number of bytes are for each plot.