Split PlayerBoolArray into two versions for players and teams
They share code and should result in the same compiled code (providing inlining is enabled)
The only difference is get/set using PlayerTypes and TeamTypes in arguments.
This means the compiler will reject code, which use both playerID and teamID as index in the same array.
Sure playerID and TeamID are usually the same, but they might not be that in custom/multiplayer games.