Updated PlayerArray to make it more generic. It now supports enum types as template variable
Each class/type has to be assigned in the header, telling the array data size and default value.
For xml enums it requires the type and NUM_*_TYPES. Both are fairly easy to set up.
Supports xml type conversion on savegame load.
Currently the only enum is PlayerTypes, which defaults to NO_PLAYER. It uses just one byte for each.

Added more PlayerArrays/TeamArrays to CvPlot to continue the conversion to CvSavegame.