EnumMap can now use CityPlotTypes for length NUM_CITY_PLOTS and can use int for index if requested
Added new enum CityPlotTypes and allow EnumMap to use it for length.
Added EnumMapInt class, which is the same as EnumMap, but it can use int for index in get/set etc
Added FOREACH macro to loop through all indexes in an xml file without typing out the entire for loop header
CvCity and CvCityAI now use EnumMap to store arrays of length NUM_CITY_PLOTS

Warning: EnumMapInt makes it easier to code because the type doesn't matter in the calls.
However this also disables compiler bug detection.
Use only if you need to typecast the arguments to all calls anyway.