EnumMap is now using static arrays instead of pointers if the array is max 4 bytes at compile time
This threshold can be set (it's an enum value), but increasing past 4 becomes a tradeoff because that will increase memory usage for empty arrays.
4 is the size of the pointer, hence no extra memory even for empty arrays.
When this feature is used, memory I/O becomes faster as there is no pointer to resolve.

Note: bools already had this feature and it can handle 8 bytes because it's optimized for one bool for each player/team.