Fix #42: escape to port help text will no longer reveal nationality of units with hidden nationality
The issue is that the help text tells which colony the unit have a chance to flee to if defeated, hence same owner.
Now the display is only always on for your own team's unit.
With no variables hidden game option:
- Revealed nationality and revealed colony: display colony name
- Show chance to free if defeated, but no colony

Anything else (hidden variables, other teams): not present in the display