Hide unused description/help/pedia/strategy strings from the user interface
If any of those 4 strings starts with TXT_KEY after being looked up, the string is cleared in the CvInfo class.
This will in most cases be detected as "intentionally left blank" in the display and not add lines to the text in question.
The language "Tag" ignores this as it is intended to display the names of the strings (translator help)

The text xml files still needs some cleaning before all the strings are displayed as they were in 2.7.x.