Translations can now be added as a subfolder instead of merged into existing files
It's now possible to add a subfolder containing a translation without also containing English.
When the same TXT_KEY is read multiple times, the game will store the last read
English and last read of the current language, even if they aren't read from the same file.

Untranslated TXT_KEYs will be written to translation.log if logging is enabled.