Added bat files to convert project files between MSVS 2010 and 2017
2017 requires extra data to unlock all IDE features (like intellisense)
Upgrading will however break support for 2010, an IDE some people still use.

Maintaining a project file for each version is bad as it contains a list of source files,
which needs to be maintained. This will go out of sync eventually.
Converting the same file between versions means only one list of source files to maintain.

Note: requires perl (like strawberry perl) to work as the bat files are frontend to a perl script.