Get revision is now in savegames. Also more makefile auto configuring paths
Compiling will now add the git revision to the DLL, which is then saved to the savegame.
It's not used, but opening a savegame with a hex editor or similar reveals which revision saved it.
(it's the second string in the file. The first is the name of the mod)
"Unknown Version" fallback for no git or no repository

Added script to detect location of compiler and git automatically.
Hopefully this will make first compilation easier without too much (if any) path setup.