EnumMap: reading a savegame with a removed xml entry is now more likely to work
Say Unit X is removed. The entry for X will now be discarded instead of trying to write to index -1.
This will make savegames more likely to survive such a change, but depending on what is removed,
other places in the savegame can become corrupted and prevent loading a working game.