#394 Fix crash caused by unit having inv. prof.
The crash itself is downstream of the failure to remove the unit
from the city, a process that is not allowed to fail according
to the current logic.

Added bForceCheck parameter to canHaveProfession to force
a recheck of the unit's profession.

Added several assert checks to add resilience in case of similar
issues in the future.