Const corrected CvSavegameWriter. This fixes a diplo related crash in hot seat games
CvSavegameWriter did not have const in arguments using references.
This meant const functions couldn't call it. As a result CvDiploParameters:write lost const.
The exe apparently requires a const write and without it, it crashed.

Why the crash apparently only happened when proposing a deal in hot seat multiplayer is unknown.