Domestic Advisor: created a commented API for DomesticAdvisorTable
Now the file is split into a public and a private part with the public on top.
The public works like a header file in C++, revealing methods exposed to the outside world.
Those methods are commented to reveal usage and all they do is call the private methods.
This way the API is not hidden inside the code, making it confusing what is API and what is internal.