#416 Fix infinite loop
CvSelectionGroupAI::AI_tradeRoutes()
A transport unit that fails to find at least a single reachable
source or destination may fail to detect this condition.
Since it later pushes a MISSION_MOVE_TO and returns true,
the path following will fail and the function  will be retried.
To fix this we return early when no candidate city is reachable