#409 Fix bug that sidestepped combat with enemies that are allowed to be captured
The feature that allowed a ship to escape a city that was lost introduced this regression.
The culprit was that canDefend() was called too early and since it checks for getCapturingPlayer() != NO_PLAYER,
this lead to the ship escape logic being invoked rather than normal combat taking place. Since the defending
ships was neither in city or fort it was always automatically destroyed by the attacker without combat.