Added "Build Release" target to the project
This will build a zip file to be uploaded on GitHub for releases.
Compiles a FinalRelease DLL if needed.
Automatically labels itself based on top version number from changelog.txt.
Relies on export-ignore to control what is included in a release.
Supports adding untracked files (used to add the dll)