Fix #401: Loading vanilla will no longer complain about a missing string due to WTP mapscript
This is done by turning our PublicMaps into PrivateMaps. THis means they are only visible to this mod.
PublicMaps are visible to other mods as well as vanilla.
The error is caused when vanilla tries to read our WTP only FaireWeather mapscript.
For whatever reason, the game will load something from the last used mapscript before reaching the main menu.

This also ensures that players will only see files intended for the current version.
Useful if we add a new file in the future and somebody has two versions installed.
The new file won't be loaded by the old version of the mod.