Fix #254: Fog of war now covers plots correctly on load
A combination of promotions extending viewing distance and weather blocking view of some plots
could result in corrupting the count of units viewing said plot when loading a savegame.
As a result it would still think a unit was watching it when all units had moved away.

The fix is to not save the view count and instead calculate it once loading is complete.
Not only will this fix the problem, it will fix savegames already broken by this issue.