Fix #442: Automated transports can get stuck doing nothing
This turned out to be caused by the fix for #416.
When transports should give up because colonies are unreachable, the same "give up" code
could make transports give up on finding somewhere to drop off cargo in certain conditions.
Specifically if the transport can't find a traderoute where the current cargo is imported as the
traderoute's source colony.