Code cleanup: TASK_YIELD_TRADEROUTE no longer uses binary operators on data transmitted on the network
Due to having one argument too little in CvCity::doTask, an int is used to transfer 2 uint16.
The code for merging and splitting isn't human readable friendly and has resulted in bugs multiple times.
This is now replaced with a class, which by usage of a union and struct can address the same
32 bit in multiple ways. For this reason the variables are accessed by name, hence removing
binary code from the code.

Technically no difference, but it's human readable and the compiler is better at catching bugs.
Hopefully this part of the code will stay bugfree this time. No known bugs this time, but previously
working code has been broken due to being fragile and hard to read.