#438 Event xml now accepts both ints and types for generic parameters
The 4 generic parameters in CvEventInfo are still ints.
Reading them can either be done using ints or a string.
If a string is used, the xml reading code will getInfoTypeForString to figure out which int to store.
Provides proper assert messages if the string from xml isn't as expected in the DLL.