MSVC project: default solution is now multi core compilation
The problem is that if no .vs folder exist, MSVC picks the first solution alphabetically.
This was Build, which will always use full program optimization and make a release zip file.
While fully optimized is good, it can take something like 15 times as long to build as standard debug.
Now less used solutions will start with z, hence be at the bottom of the list.