Fix #449: AI units no longer get stuck on large rivers
Previous fix didn't work in all cases. This is a fundamentally different approach.
Instead of spot fixing, this alters area() and getArea() in CvUnit and CvPlot.
CvUnit now always assume it's on a land plot next to the river if it's on a large river.
This is true even if the function calling CvUnit::area() and getArea() isn't updated.

Calls to area and getArea in CvPlot needs updating for this to activate.