WidgetTypes are now back in vanilla order
The exe has some widget types hardcoded and it can result in weird bugs if the DLL has different values.
Now the order is restored and static asserts prevents the order from going out of sync again.
We are free to add new widget types after the vanilla ones as long as we do not change the vanilla values.