Fix: raiding natives can no longer cause negative yields stored in cities
The issue occurred when natives would change profession as a result of the raid.
When checking if it's possible, yield cost modifiers were not considered.
This meant modifiers from traits, handicap and game speed could make the code test against the wrong value
Now the cached yield cost in CvPlayer is used both for check and for changing profession.

Also removed a hack, which added yields to the city to try to avoid negative numbers (without success).