Fix: raw format keywords no longer showing up in popup window checkbox help text
Apparently the is displaying raw text in that help text and not passing formatting.
Made the DLL remove the formatting prior to sending it to the exe.

Note: looks like an exe bug, but I have written 100% of the code using that help popup.
In other words it's an exe bug, but not vanilla bug as vanilla doesn't use this feature.