Fix: events can no longer result in colonies with negative food stored
Calculating food change is now done in a function in CvEvent to make the change consistent.
Now both testing and applying takes game speed into account. This was the culprit.
On Marathon, it could test for -50 and apply -150.

Applying food change to all colonies is now capped at 0, meaning it can't go negative.