Fix: AI/automated ships no longer stop due to viewing land units as dangerous
Automated units looked at the same area as they are in for danger. This meant units in cities looked at
land. Since the land units would not move away, the ship would get stuck, just waiting for
hostile land units to move into the city. This is particularly true when they got stuck in native cities.

This also makes land units look at the land area for danger even if standing on a large river.