Fix #471 Fixed spacing issue in yield help popup in the trade window
The issue was that the line contained the trade port followed by spaces and then the trading price.
The spaces were two tabs and since "Port Royal" was so much longer than the other two
it acted like it had an extra tab.

Also changed German to have the same spacing as the other languages.