Fix #472 Natives will no longer sell more cargo than they have in their settlement (negative storage)
The issue seems to be that the natives can offer something to the player and sent a diplo event.
When the human player starts his/her turn, the offer is displayed and the player accepts.
If the stored amount in the native settlement is reduced in the meantime, negative storage occurred.
Now the code will set storage to 0 if this happens.