Fix #399: Pressing F10 no longer triggers an assertion failure
The issue was that CONTROL_ACHIEVE_SCREEN was not in the switch case in the DLL.
However the feature works as it is handled by CvEventManager.py.
This mean the assertion is correct in assuming it's a high risk case for a bug,
but in this specific case it wasn't a bug.