InfoArray can now use a static array if it only needs to store max 2 variables
The pointer to the allocated array takes up 4 bytes. Each variable is only 2 bytes.
If the array can fit in the memory for the pointer, the pointer itself becomes the array.
This lowers memory latency when reading the array.
It also obviously lowers memory usage and avoids the slow memory allocations, but that's minor.

Also switched to using memcpy when copying arrays. This triggers hardware acceleration of memory copying.