InfoArray replaced getWithTemplate with get0..get3
InfoArray had the issue that getUnit, getBonus etc works well if there is only one of each.
getInt for InfoArray<UnitTypes, int, int> will fail to compile because it has 2 different ints.
Now that case can be solved with get1 and get2.
get+number is a fallback for templates and for multiple parameters of the same type.
If there is just one, use the name as it results in more readable code.