Made some adjustment for scouting behavior
Land danger functions now ignore animals (which are generally to weak to be considered a credible threat). Note that this was already implemented for water animals.
Experts are now considered less than ideal for the scout profession which should have the effect that free colonists etc. will be preferred over them
Increased the utility of a city emitting a scout in case the area has none
Outdoor professions will now attempt to speak to the chief if they happen to be in a native city