Compiler now enforce newline at the end of files
Recommended approach to git/svn/etc is to end files with a newline.
Failure to do so will mark the last line as modified if more lines are added later, messing up history.
The compiler now causes an error if last character in a file isn't a newline.

This will not change the compiled file at all, but it will make git history more clean.
It's also now impossible to forget to check for this issue prior to committing.