EnumMap can now be a list of classes
If type isn't of a known type (int, enum etc) then this mode will be used.
Get/set can't be used. Instead [] access is required. This makes more sense when accessing classes.

Code cleanup is needed as the current implementation can conflict with the old EnumMap code.
This is detected by a failure to compile, hence fairly serious if it triggers.