Removed EnumMap2D as it is replaced with EnumMap of EnumMaps
EnumMap2D was an EnumMap like class, which contained an array of EnumMaps.
Now that EnumMap can contain an array of class instances, an EnumMap of EnumMaps is to be preferred.
EnumMap[index1]. can now be used to access all EnumMap functions rather than the subset defined in EM2D.
Since EnumMap can handle classes, it can handle EnumMap of EnumMaps of... there is no hard limit.

Fixed the known bugs in EnumMap of classes.
Added savegame support to EnumMap of classes.