Global variables storing length of xml files are now read only
Constants like NUM_UNIT_TYPES are declared as enum values when xml values are hardcoded.
When xml data is set at runtime, the same names are used for global variable, effectively allowing usage
of constants like NUM_UNIT_TYPES to always be present.
This update makes the constant read only except in CvGlobalsEnumSetup.cpp
This way if code tries to alter the length of an xml file by accident the compiler will throw an error.
Examples of such a mistake would be typing = instead of ==