Domestic Advisor: domestic market yields are no longer hardcoded
The yields used by units or buildings are stored in an InfoArray, which is generated at startup.
This saves looping in CvCity::doTurn and allows the DA to dynamically set the columns needed.
Update xml to use a new type of yield and the new column will appear.

To support this, the DA "engine" now supports setting columns based on any InfoArray.
Currently only works on types with getChar support, though other rules for header text can be added.

Note: while the new page for the DA exist in code, it's currently not used because new and old style
pages can't coexist under the same button. This will be fixed once all pages use the new system.