Domestic Advisor: A button can now contain multiple pages
The system states one source file for each page. One page can be displayed as multiple pages, like
if there is one column for each yield. It's still only one page and one rule set to fill cells.
Adding buttons in the DA window is done by appending a function call, which links the button to a page.

This update adds a page, which contains a list of pages. This way a button can contain multiple pages,
each with a unique source file with unique code for setting up columns and filling cells.

Both places to add a page works by appending a page source file to an array. While the syntax in each
differs, they are both one line additions. It's intentionally designed to share the page source files.
As a result page files can freely be moved between multipage buttons or get a button of their own.