Domestic Advisor: added teacher advisor (Raubwürger)
New advisor page listing which experts can be taught at schools in the colonies.
This provides an easy to view overview of not only which colonies have the needed schools, but also
which colonies have the needed profession in the school already.

Originally written by Raubwürger. Converted to new DA file format and removed hardcoding by Nightinggale.

Note: conversion isn't complete yet. Two columns are missing and the headers are blank.