Fixed crash in CvPlayer::redistributeWood()
getCity(0) might return NULL and this wasn't checked.
If a list contains 3 cities and the first one dies, then getCity() will return NULL for all but 1 and 2.
Index 0 will be used for the next city added to the list meaning we can't assume all indexes to be used.
Also we can't assume the first index to be the oldest city.

For this reason always use firstCity and nextCity as they skip invalid IDs.