Fix: starting a scenario will no longer crash if it contains a team with no players in it
This issue caused Americas Gigantic to crash because the barbarian player had been removed from the file.
However other teams still has the memory of being at war with the barbarian team, which caused a crash
when setting up diplomatic relationships.

Now empty players as well as teams with no players in them will ignore all data when loading.
This includes when team A sets up relationship with team B and team B is empty.

The barbarian player is (currently) not supposed to be in scenarios as it is automatically added
each time a new game is started, be it scenario or random map.