Added a bat file to download/install the compiler
The script is Get Compiler.bat in project files.
Double clicking this will clone or pull the compiler meaning it will also update if needed.
It will then copy tbb.dll and tbbmalloc.dll to Colonization.exe if needed.
Last it will open the strawberry perl download page if it isn't already installed.