Fix: project file folder navigation now works again
The project .filter file (xml) lacked a tag closing, preventing it from being read.
The result was that all files were displayed in the same location instead of being sorted.
We are now back to having for python interface, Cv files, data container classes etc.
Also added the KmodPathFinder.cpp/h files to the project file.

None of this affects the compiled DLL. It's purely for navigating files inside MSVC.