Compiling: autogenerated header files no longer generate temp files
New headers were written as temp files and only renamed if updated. This way the file modification date
won't update without content updating. Now the new files is written in memory and only written to the
disk when it differs from the existing files. Modification dates are still used by the compiler to
determine what needs to be compiled again.

This will result in a performance boost when nothing needs to be compiled, which happens when you start
the mod from within MSVC (like with a debugger attached).