Compile.pl: added the ability to create a bat file to compile and start the game
The menu in compile.pl now offers the option to create a bat file, which works like an alias to start the
mod. In other words it can be moved anywhere on the computer and it will still work.
Unlike the alias solution, the bat file will compile the DLL file first, ensuring it's always up to date.
The makefile will only compile outdated code meaning it's done in a few seconds if nothing changed.

Removed Compile-Assert.bat and Compile-Release.bat as compile.pl now does the job of both and more.