The compiler is now aware of GlobalDefines.xml values
Since we know vanilla won't change, hardcode the values into the DLL.
If any of the tags are present in GlobalDefinesALT, then use the modded values instead.

This will benefit debugger usability as well as potentially aiding runtime performance.
The only tradeoff is that adding new vanilla tags to GlobalDefinesALT requires recompiling the DLL.