The enum strings in GlobalTypes.xml are now available inside the DLL
While currently unused, this can also be used to generate enums, eliminating the risk of xml going out
of sync with the DLL. I didn't make use of that because that has already happened and needs cleanup.

This will eventually turn into an automated system where the type strings can be used as values
in other xml files just like the Type string is used.