#561: New pathfinding now predicts how many turns it takes to move to a certain plot
It now reports how many turns it will take before a unit is ready to move away from a plot.
This doesn't change the pathfinder and it will not always pick the fastest path.

USE_CLASSIC_MOVEMENT_SYSTEM is now always hardcoded and used by the precompiler.
It's still set in globalDefineALT, but changing it requires recompiling the DLL.