World size is no longer hardcoded in the DLL
To accomplish this, the python interface setup introduced the new concept of looping xml files while
assigning the values of the enum. This way the enum values will not have to be known at compile time.

CvPlayerAI::AI_foundValueNative now does runtime string check for gigantic maps.
Ideally this can be done in a more clean way, but it works and was the last thing preventing the DLL
from accepting xml changes without recompiling.