Fix #749: game option to remove world builder will now always do so in release builds
It is still possible to get the game to ignore this game option, but now it will only be ignored
in DEBUG and ASSERT builds when cheats are enabled as that would be the setup to explore bugs.
Release builds can no longer enable cheats to enable world builder when disabled by this option.