#771 The global jobMutex has been replaced by a per-player mutex
Note that we are not likely to experience any difference in performance since we currently process each player cities sequentially. However, If we were to allow all cities in the game to be processed in parallel then we'd likely see significant gains though

#771 CvGame::calculateSyncChecksum now includes a hash of each city's citizens

#771 The job assigner does not allow bumping slot workers inside the (innermost) parallel loop

Instead we let the juggling take care of any bumping since that avoids race-conditions

#771 Refactored AI_juggleColonist

Renamed AI_professionValue to AI_citizenProfessionValue

#771 CvCityAI::AI_juggleColonist now ignores colonists with NO_PROFESSION

Thus we avoid an assertion since we may occasionally end up with unemployed citizens (i.e. when running out of viable plots / slots to assign the citizen(s))