Added an error popup when starting a new game if the map lacks clay or there aren't any AI players
Wetlands (hence clay) is missing with outdated map scripts, hence why it would be good to tell people
if they somehow ended up with a map, which isn't intended to be playable.

When playing multiplayer, the European player slots are open by default to allow new players.
If a slot is still open when starting the game, no AI will take the place, leaving the risk of
having a game with two humans and no AI players. That is often not the intension.