Adding a vector for captured unit IDs to hopefully prevent cargo from being destroyed while being captured
The problem is that killing (or giving to a new owner) a transport will instantly kill the cargo.
This will capture cargo prior to transports (hopefully every time).
What this does for sure is to not mess up looping through all the units in question.
It also adds NULL check to ensure another failure will just skip capturing a unit rather than crashing.