#696: added wrapper classes to avoid exe and python calling our code directly
Not calling the Cv headers directly gives us freedom to change the functions without messing up the
interface to either the exe or python. Even better a broken interface becomes a compile time error.

Altered the DllExport check to report better on errors and generally work better with the new setup.

The individual functions have to be added individually. This commit will mainly just add the classes
to contain the new wrapper functions as well as surrounding code to make it just work once they are there.