Fix #808: yield production bonuses from traits are now applied to all cities
The previous approach applied to city plots if that plot already produced at least one of that yield.
Even worse, city plots can only produce one cargo yield other than food so it likely wouldn't be used.
Now it is applied as a new item on the help popup called "from Leader" and is always applied.