EXE_interface.cpp: DllExport with incorrect indentation will no longer cause weird compile errors
DllExport testing needs to be picky because any mistake can cause the exe to crash.
Now the word DllExport is banned in this file unless it is correctly picked up during verification.
This prevents a later error, which is usually in a header with no description to what actually went wrong.