#807: Length of xml files are now set as soon as the exe attaches to the dll
A weakness of EnumMap is that it breaks hard if the constructor is unaware of the xml length.
Vanilla doesn't know the length prior to calling constructors of multiple classes, including CvGlobals.
With this change, EnumMap will work as members of any class or in any function.

Note: incomplete change. Not all xml files have their length set this early.

As a bonus we now have the full tinyxml2 library and not just the vanilla subset.